Add files via upload
Build a Linear Machine Learning model to understand the relationship between the population of US cities in the year 1920 and 1930. Evaluate the model performance with appropriate measures. Perform all the required graphical and quantitative exploratory data analysis prior to model building. The dataset has 49 rows and 2 columns. The measurements are the population (in 1000's) of 49 U.S. cities in 1920 and 1930. The 49 cities are a random sample taken from the 196 largest cities in U.S